Moving kotlin snipets to kotlin directory in docs module
