+++ stream-chat-android-docs/src/main/kotlin/io/getstream/chat/docs/kotlin/ui/guides/realm/repository/RealmAttachmentRepository.kt
@@ -1,0 +1,47 @@
+/*
+ * Copyright (c) 2014-2022 Stream.io Inc. All rights reserved.
+ *
+ * Licensed under the Stream License;
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    https://github.com/GetStream/stream-chat-android/blob/main/LICENSE
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.getstream.chat.docs.kotlin.ui.guides.realm.repository
+
+import io.getstream.chat.android.client.persistance.repository.AttachmentRepository
+import io.getstream.chat.android.models.Attachment
+import io.getstream.chat.docs.kotlin.ui.guides.realm.entities.AttachmentEntityRealm
+import io.getstream.chat.docs.kotlin.ui.guides.realm.entities.toDomain
+import io.realm.kotlin.Realm
+import io.realm.kotlin.ext.query
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.map
+
+internal class RealmAttachmentRepository(private val realm: Realm) : AttachmentRepository {
+
+    override fun observeAttachmentsForMessage(messageId: String): Flow<List<Attachment>> {
+        return realm.query<AttachmentEntityRealm>("message_id == '$messageId'")
+            .asFlow()
+            .map { results ->
+                results.list.map { attachmentEntity ->
+                    attachmentEntity.toDomain()
+                }
+            }
+    }
+
+    override suspend fun clear() {
+        val allAttachments = realm.query<AttachmentEntityRealm>().find()
+
+        realm.write {
+            delete(allAttachments)
+        }
+    }
+}
